backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.2
python-decouple>=3.8
Pillow>=10.0
requests>=2.31
djangorestframework-simplejwt>=5.3
